pyproject.toml
[project]
name = "chatbot-with-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
